requirements.txt
holidays==0.13
numpy==1.22.1
pandas==1.4.0
pyparsing==3.0.7
python_dateutil==2.8.2
